server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
thefuzz
Pillow
pytesseract
google-generativeai
pydantic
pydantic-settings
python-dotenv
aiofiles
firebase-admin
opencv-python-headless
python-Levenshtein